refactor(preventive): corrigido tipo de dado para o atributo defaultFrequencyDays

+++ api/src/main/java/com/ong/acolhepatinhas/api/veterinary/preventive/catalog/DTO/NewPreventiveProcedureRequest.java
@@ -13,6 +13,6 @@ public record NewPreventiveProcedureRequest(
 
     @Positive
     @Schema(example = "180")
-    Integer defaultFrequencyDays
+    short defaultFrequencyDays
 ) {
 }
