+++ api/src/main/java/com/ong/acolhepatinhas/api/veterinary/preventive/catalog/PreventiveProcedure.java
@@ -26,5 +26,5 @@ public class PreventiveProcedure {
 
     private String name;
 
-    private Integer defaultFrequencyDays;
+    private short defaultFrequencyDays;
 }
